// src/enums/contants.notifications.ts
export enum NOTIFICATIONS {
    EMAIL_PASSWORD_INCORRECT = "Email o Contraseña incorrecta.",
    PRODUCT_CREATED = "Producto agregado correctamente.",
    PRODUCT_EXIST = "El nombre del producto ya está registrado.",
    LOGIN_SUCCESS = "¡Inicio de sesión exitoso!",
    LOGIN_ERROR_GENERIC = "Error al iniciar sesión. Por favor, verifica tus credenciales.",
    LOGIN_ERROR_UNEXPECTED = "Ocurrió un error inesperado al intentar iniciar sesión. Inténtalo más tarde.",
    ACCOUNT_LOCKED = "Tu cuenta ha sido bloqueada temporalmente. Intenta de nuevo más tarde o contacta a soporte.",
    ACCOUNT_DISABLED = "Tu cuenta ha sido deshabilitada. Por favor, contacta a soporte.",
    INVALID_CREDENTIALS = "Email o contraseña incorrectos.",
}