// src/enums/constants.login.ts
export enum LOGIN_TEXT {
    EMAIL = "Email:",
    USERNAME = "Nombre de usuario:",
    PASSWORD = "Contraseña:",
    LOGIN = "Iniciar Sesión",
    FORGOT_PASSWORD = "¿Olvidaste tu contraseña?",
    REMEMBER = "Recordar",
    REQUIRED_USERNAME = "Ingrese un nombre de usuario",
    REQUIRED_PASSWORD = "Ingrese una contraseña valida",
    REQUIRED_EMAIL = "Ingrese un email valido",
}